+++ NEWS.md
@@ -1,6 +1,7 @@
-# targets 1.10.1.9002 (development)
+# targets 1.10.1.9003 (development)
 
 * Upload workspaces to the cloud if `tar_option_get("repository_meta")` is `"aws"` or `"gcp"`. Download them with `tar_workspace_download()` and delete them with `tar_destroy(destroy = "all")` or `tar_destroy(destroy = "cloud")`.
+* Deep-copy settings when resolving `format = "auto"` (#1425, @paulseamer).
 
 # targets 1.10.1